Move methods from API to TEBackpack
Also cleanup them somewhat
